Policies (error, optional, parse) moved to different header files for the sake of modularity
